+++ source/engine/system/TimeSystem.h
@@ -14,8 +14,8 @@ public:
 	void shutdown() override;
 
 	const time_t getGameStartTime() const override;
-	const double getDeltaTime() const override;
-	const double getcurrentTime() const override;
+	const long long getDeltaTime() const override;
+	const long long getcurrentTime() const override;
 	const std::tuple<int, unsigned, unsigned> getCivilFromDays(int z) const override;
 	const std::string getCurrentTimeInLocal(std::chrono::hours timezone_adjustment = std::chrono::hours(8)) const override;
 	const std::string getCurrentTimeInLocalForOutput(std::chrono::hours timezone_adjustment = std::chrono::hours(8)) const override;
